Refactor inner classes into separate files for Java conventions
Extract InitialSnapshotCache, ScanFileRDD, SerializableFileStatus,
and SerializableScanData into their own top-level Java files.
